refactor: removed React support instead using SSR template by Qute
