--- src/main/java/id/putra/peoplelibrarysystem/controller/AdminController.java
@@ -1,5 +1,0 @@
-package id.putra.peoplelibrarysystem.controller;
-
-public class AdminController {
-
-}
