+++ src/main/resources/application.properties
@@ -11,3 +11,5 @@ smallrye.jwt.sign.key.location=${PRIVATE_KEY_LOCATION}
 
 mp.jwt.token.cookie=jwt
 mp.jwt.token.header=Cookie
+
+quarkus.qute.property-not-found-strategy=default
