--- src/main/webui/src/components/ProfileComponent.tsx
@@ -1,22 +1,0 @@
-import { Menu, MenuItem } from "@mui/material";
-import useProfileStore from "../store/profileStore";
-
-export default function ProfileComponent() {
-  const { anchorEl, open, handleClose } = useProfileStore();
-
-  return (
-    <Menu
-      id="basic-menu"
-      anchorEl={anchorEl}
-      open={open}
-      onClose={handleClose}
-      MenuListProps={{
-        "aria-labelledby": "basic-button",
-      }}
-    >
-      <MenuItem onClick={handleClose}>Profile</MenuItem>
-      <MenuItem onClick={handleClose}>My account</MenuItem>
-      <MenuItem onClick={handleClose}>Logout</MenuItem>
-    </Menu>
-  );
-}
