--- src/main/webui/src/main.tsx
@@ -1,21 +1,0 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import { createRouter, RouterProvider } from "@tanstack/react-router";
-import { routeTree } from './routeTree.gen.ts'
-import CssBaseline from '@mui/material/CssBaseline';
-
-
-const router = createRouter({ routeTree })
-
-declare module '@tanstack/react-router' {
-  interface Register {
-    router: typeof router
-  }
-}
-
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <RouterProvider router={router} />
-    <CssBaseline />
-  </StrictMode>,
-)
