--- src/main/webui/src/store/drawerStore.ts
@@ -1,17 +1,0 @@
-import { create } from "zustand";
-
-interface DrawerState {
-  open: boolean;
-  toggleDrawer: (newOpen: boolean) => () => void;
-}
-
-const useDrawerStore = create<DrawerState>((set) => ({
-  open: false,
-  toggleDrawer: (newOpen: boolean) => () => {
-    set({
-      open: newOpen,
-    });
-  },
-}));
-
-export default useDrawerStore;
