--- src/main/webui/src/store/profileStore.ts
@@ -1,22 +1,0 @@
-import { create } from "zustand";
-
-interface ProfileState {
-  open: boolean;
-  anchorEl: HTMLElement | null;
-  handleClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
-  handleClose: VoidFunction;
-}
-
-const useProfileStore = create<ProfileState>((set) => ({
-  open: false,
-  anchorEl: null,
-  handleClick: function (event: React.MouseEvent<HTMLButtonElement>) {
-    set({
-      anchorEl: event.currentTarget,
-      open: Boolean(event.currentTarget),
-    });
-  },
-  handleClose: () => set({ open: false, anchorEl: null }),
-}));
-
-export default useProfileStore;
