--- src/main/webui/vite.config.ts
@@ -1,13 +1,0 @@
-import {defineConfig} from 'vite'
-import react from '@vitejs/plugin-react-swc'
-import TanStackRouterVite from "@tanstack/router-plugin/vite";
-
-// https://vite.dev/config/
-export default defineConfig({
-    plugins: [
-        react(),
-        TanStackRouterVite({
-            autoCodeSplitting: true
-        }),
-    ],
-})
